+++ spec/punchblock/translator/freeswitch/component/record_spec.rb
@@ -112,11 +112,10 @@ module Punchblock
 
               context "set to a positive number" do
                 let(:command_options) { { :initial_timeout => 10 } }
-                it "should return an error and not execute any actions" do
-                  mock_call.should_receive(:uuid_foo).never
+                it "should execute normally" do
+                  mock_call.should_receive(:uuid_foo).once.with(:record, /.wav$/)
                   subject.execute
-                  error = ProtocolError.new.setup 'option error', 'An initial-timeout value is unsupported.'
-                  original_command.response(0.1).should be == error
+                  original_command.response(0.1).should be_a Ref
                 end
               end
             end
@@ -142,11 +141,10 @@ module Punchblock
 
               context "set to a positive number" do
                 let(:command_options) { { :final_timeout => 10 } }
-                it "should return an error and not execute any actions" do
-                  mock_call.should_receive(:uuid_foo).never
+                it "should execute normally" do
+                  mock_call.should_receive(:uuid_foo).once.with(:record, /.wav$/)
                   subject.execute
-                  error = ProtocolError.new.setup 'option error', 'A final-timeout value is unsupported.'
-                  original_command.response(0.1).should be == error
+                  original_command.response(0.1).should be_a Ref
                 end
               end
             end
@@ -298,6 +296,48 @@ module Punchblock
                 let(:command_options) { { :direction => :recv } }
                 it "should execute the setvar application with recv options before recording" do
                   mock_call.should_receive(:uuid_foo).once.with(:setvar, "RECORD_READ_ONLY true").ordered
+                  mock_call.should_receive(:uuid_foo).once.with(:record, /.wav$/).ordered
+                  subject.execute
+                  original_command.response(0.1).should be_a Ref
+                end
+              end
+            end
+
+            describe 'initial-timeout' do
+              context "when not set or nil" do
+                let (:command_options) { { :initial_timeout => nil} }
+                it "should setvar RECORD_INITIAL_TIMEOUT_MS with a 0 value" do
+                  mock_call.should_receive(:uuid_foo).once.with(:setvar, "RECORD_INITIAL_TIMEOUT_MS 0").ordered
+                  mock_call.should_receive(:uuid_foo).once.with(:record, /.wav$/).ordered
+                  subject.execute
+                  original_command.response(0.1).should be_a Ref
+                end
+              end
+              context "when set to a ms value" do
+                let (:command_options) { { :initial_timeout => 5000} }
+                it "should setvar RECORD_INITIAL_TIMEOUT_MS with a value in ms" do
+                  mock_call.should_receive(:uuid_foo).once.with(:setvar, "RECORD_INITIAL_TIMEOUT_MS 5000").ordered
+                  mock_call.should_receive(:uuid_foo).once.with(:record, /.wav$/).ordered
+                  subject.execute
+                  original_command.response(0.1).should be_a Ref
+                end
+              end
+            end
+
+            describe 'final-timeout' do
+              context "when not set or nil" do
+                let (:command_options) { { :final_timeout => nil} }
+                it "should setvar RECORD_FINAL_TIMEOUT_MS with a 0 value" do
+                  mock_call.should_receive(:uuid_foo).once.with(:setvar, "RECORD_FINAL_TIMEOUT_MS 0").ordered
+                  mock_call.should_receive(:uuid_foo).once.with(:record, /.wav$/).ordered
+                  subject.execute
+                  original_command.response(0.1).should be_a Ref
+                end
+              end
+              context "when set to a ms value" do
+                let (:command_options) { { :final_timeout => 5000} }
+                it "should setvar RECORD_FINAL_TIMEOUT_MS with a value in ms" do
+                  mock_call.should_receive(:uuid_foo).once.with(:setvar, "RECORD_FINAL_TIMEOUT_MS 5000").ordered
                   mock_call.should_receive(:uuid_foo).once.with(:record, /.wav$/).ordered
                   subject.execute
                   original_command.response(0.1).should be_a Ref